TypeScriptified "fetch email analytics Mailgun events" utility
ref https://github.com/TryGhost/Ghost/pull/29845

This change should have no user impact.
